# api/serve.openapi.yaml
openapi: 3.1.0
info:
  title: openstash serve API
  version: 0.4.0
  description: HTTP API over the local openstash OpenAPI cache.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/specs/{specKey}/versions/{version}/operations:
    get:
      summary: Search, show, or gather operations
      operationId: operations
      parameters:
        - $ref: "#/components/parameters/specKey"
        - $ref: "#/components/parameters/version"
        - name: detail
          in: query
          schema:
            type: string
            enum: [search, show, gather]
            default: search
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
        - name: pathPrefix
          in: query
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 0
        - name: expand
          in: query
          schema:
            type: boolean
            default: false
        - name: in
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [paths, schemas]
      responses:
        "200":
          description: Operation results
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    specKey:
      name: specKey
      in: path
      required: true
      schema:
        type: string
    version:
      name: version
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Spec or operation not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...